requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain-google-genai
gradio_client
pydantic
requests
opencv-python-headless
google-genai
python-multipart
Pillow
groq
numpy
gradio
google-generativeai

google-cloud-aiplatform

google-cloud-storage
